+++ src/pages/jeongeun/Main/Comments.js
@@ -1,8 +1,6 @@
 import React from 'react';
 
 const Comments = (props) => {
-  console.log(props);
-
   return (
     <>
       <span className="commentList">
