+++ src/pages/jeongeun/Main/Main.scss
@@ -89,6 +89,7 @@
         width: 470px;
         height: 450px;
         margin-right: 30px;
+        margin-bottom: 10px;
 
         article {
           position: relative;
@@ -119,33 +120,36 @@
               padding: 5px;
             }
           }
-
-          .commentbox {
+          .commentBigBox {
             display: flex;
-            justify-content: space-between;
-            width: 100%;
-            overflow-x: hidden;
+            margin: 0;
+            .commentbox {
+              display: flex;
+              justify-content: space-between;
+              width: 100%;
+              overflow-x: hidden;
 
-            .allcomment {
-              list-style: none;
-              padding-left: 10px;
-              text-align: left;
-              font-size: 12px;
+              .allcomment {
+                list-style: none;
+                padding-left: 10px;
+                text-align: left;
+                font-size: 12px;
 
-              .newComment {
-                postion: relative;
-                width: 100%;
-              }
+                .newComment {
+                  postion: relative;
+                  width: 100%;
+                }
 
-              .removeBtn {
-                position: absolute;
-                right: 10px;
-                align-self: flex-end;
-                margin: 0 10px;
-                color: gray;
-                background-color: white;
-                border: none;
-                font-size: 8px;
+                .removeBtn {
+                  position: absolute;
+                  right: 10px;
+                  align-self: flex-end;
+                  margin: 0 10px;
+                  color: gray;
+                  background-color: white;
+                  border: none;
+                  font-size: 8px;
+                }
               }
             }
           }
